+++ examples/javaScript/sell/inventory/createOrReplaceInventoryItem.js
@@ -4,9 +4,11 @@ var clientId = process.env.EBAY_CLIENT_ID || 'YOUR_KEY';
 var clientSecret = process.env.EBAY_CLIENT_SECRET || 'YOUR_SECRET';
 
 var eBay = require('../../../../lib/eBay-node-client')(clientId, clientSecret);
+var utils = require('../../../javaScript/utils');
 
-var browseRequest = function () {
-    var userToken = 'USER_TOKEN';
+var inventoryRequest = function () {
+    
+    var userToken = utils.USER_TOKEN;
     eBay.setUserToken(userToken);
 
     var data = {
@@ -47,4 +49,4 @@ var browseRequest = function () {
     });
 };
 
-browseRequest();
+inventoryRequest();