+++ examples/javaScript/sell/inventory/getInventoryItems.js
@@ -4,10 +4,11 @@ var clientId = process.env.EBAY_CLIENT_ID || 'YOUR_KEY';
 var clientSecret = process.env.EBAY_CLIENT_SECRET || 'YOUR_SECRET';
 
 var eBay = require('../../../../lib/eBay-node-client')(clientId, clientSecret);
+var utils = require('../../../javaScript/utils');
 
 var inventoryRequest = async function () {
 
-    var userToken = 'USER_TOKEN';
+    var userToken = utils.USER_TOKEN;
     eBay.setUserToken(userToken);
 
     var data = {