+++ examples/javaScript/utils/index.js
@@ -1,0 +1,4 @@
+'use strict';
+module.exports = {
+    USER_TOKEN: process.env.EBAY_USER_TOKEN,
+};